refactor(contracts): move feature-internal DTOs into contracts (Q-107)
Relocate CodeSnippetPersistenceDTO, UpcomingReminderDTO,
TaskTemplateHistory{Server,Client}DTO and NotificationHistoryServerDTO
into @dailyuse/contracts, and reconcile the repository Folder DTOs onto
the canonical contracts types (add FolderServerDTO, drop the divergent
local copies). Entity to*DTO() now returns the contracts types directly,
removing the `as unknown as` casts in the folder use-cases.
