+++ packages/contracts/src/modules/governance/value-objects/code-snippet.ts
@@ -16,3 +16,20 @@ export interface CodeSnippetDTO {
   type: SnippetType;
   caption: string | null;
 }
+
+// ============ Persistence DTO (持久化层) ============
+
+/**
+ * CodeSnippet Persistence DTO — database storage format.
+ * 代码片段持久化 DTO — 数据库存储格式。
+ *
+ * @internal Repository implementation detail. Consumers should use CodeSnippetDTO.
+ * @internal 仓储实现细节，消费者应使用 CodeSnippetDTO。
+ */
+export interface CodeSnippetPersistenceDTO {
+  id: string;
+  language: string;
+  content: string;
+  type: string;
+  caption: string | null;
+}
