+++ packages/contracts/src/modules/governance/value-objects/index.ts
@@ -10,4 +10,6 @@ export { Language } from './language';
 export { SnippetType } from './snippet-type';
 export { ChangeType } from './change-type';
 export type { CodeSnippetDTO } from './code-snippet';
+/** @internal Persistence format — use CodeSnippetDTO instead. 持久化格式 — 请使用 CodeSnippetDTO。 */
+export type { CodeSnippetPersistenceDTO } from './code-snippet';
 export type { RuleTagDTO } from './rule-tag';
