+++ packages/contracts/src/modules/task/entities/index.ts
@@ -8,3 +8,11 @@ export type {
 export type {
   SubtaskClientDTO,
 } from './subtask-client';
+
+export type {
+  TaskTemplateHistoryServerDTO,
+} from './task-template-history-server';
+
+export type {
+  TaskTemplateHistoryClientDTO,
+} from './task-template-history-client';
