+++ packages/contracts/src/modules/task/entities/task-template-history-client.ts
@@ -1,0 +1,7 @@
+export interface TaskTemplateHistoryClientDTO {
+  id: string;
+  templateId: string;
+  action: string;
+  changes: unknown;
+  createdAt: number;
+}
