+++ packages/contracts/src/modules/task/entities/task-template-history-server.ts
@@ -1,0 +1,7 @@
+export interface TaskTemplateHistoryServerDTO {
+  id: string;
+  templateId: string;
+  action: string;
+  changes: unknown;
+  createdAt: number;
+}
