+++ packages/governance/src/server/domain/value-objects/code-snippet.ts
@@ -15,23 +15,18 @@ import type { Result } from '@dailyuse/contracts/result';
 import { ok, error } from '@dailyuse/contracts/result';
 import type { CodeSnippetDTO } from '@dailyuse/contracts/governance';
 import type { CodeSnippetId } from '@dailyuse/contracts/governance';
+import type { CodeSnippetPersistenceDTO } from '@dailyuse/contracts/governance';
 
 // ============ Persistence DTO (持久化层) ============
 
 /**
- * CodeSnippet Persistence DTO — database storage format.
- * 代码片段持久化 DTO — 数据库存储格式。
+ * CodeSnippet Persistence DTO — re-exported from contracts.
+ * 代码片段持久化 DTO — 从 contracts 重新导出。
  *
  * @internal Repository implementation detail. Consumers should use CodeSnippetDTO.
  * @internal 仓储实现细节，消费者应使用 CodeSnippetDTO。
  */
-export interface CodeSnippetPersistenceDTO {
-  id: string;
-  language: string;
-  content: string;
-  type: string;
-  caption: string | null;
-}
+export type { CodeSnippetPersistenceDTO };
 import { Language, type Language as LanguageValue } from './language';
 import { SnippetType, type SnippetType as SnippetTypeValue } from './snippet-type';
