+++ packages/repository/src/server/application/use-cases/commands/move-folder.use-case.ts
@@ -112,6 +112,6 @@ export class MoveFolderUseCase {
 
     await this.updateResourcesRecursively(String(folder.id), folder.path);
 
-    return ok({ folder: folder.toClientDTO() as unknown as FolderClientDTO });
+    return ok({ folder: folder.toClientDTO() });
   }
 }
